Module_4/requirements.txt
flask
psycopg2-binary
pytest
pytest-cov
sphinx
sphinx-rtd-theme